bulkpb: move backuppb.IngestionPerformanceStats out of ccl
This message was relied on by kv code. Nobody used the message at its old path.
It's just used for trace events. This PR just moves its code to
`kv/bulk/bulkpb` from `ccl/backupccl/backuppb`.

Part of #91714

Epic: none

Release note: None
