+++ pkg/BUILD.bazel
@@ -1105,6 +1105,7 @@ GO_TARGETS = [
     "//pkg/keys:keys",
     "//pkg/keys:keys_test",
     "//pkg/keysbase:keysbase",
+    "//pkg/kv/bulk/bulkpb:bulkpb",
     "//pkg/kv/bulk:bulk",
     "//pkg/kv/bulk:bulk_test",
     "//pkg/kv/kvbase:kvbase",
@@ -2481,6 +2482,7 @@ GET_X_DATA_TARGETS = [
     "//pkg/keysbase:get_x_data",
     "//pkg/kv:get_x_data",
     "//pkg/kv/bulk:get_x_data",
+    "//pkg/kv/bulk/bulkpb:get_x_data",
     "//pkg/kv/kvbase:get_x_data",
     "//pkg/kv/kvclient:get_x_data",
     "//pkg/kv/kvclient/kvcoord:get_x_data",
