+++ pkg/ccl/backupccl/backuppb/BUILD.bazel
@@ -44,17 +44,13 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/cloud",
-        "//pkg/roachpb",
         "//pkg/sql/catalog/descpb",
         "//pkg/sql/parser",
         "//pkg/sql/protoreflect",
         "//pkg/sql/sem/tree",
         "//pkg/util/bulk",
-        "//pkg/util/humanizeutil",
-        "//pkg/util/log",
         "//pkg/util/uuid",
         "@com_github_cockroachdb_errors//:errors",
-        "@com_github_cockroachdb_redact//:redact",
         "@com_github_gogo_protobuf//jsonpb",
         "@io_opentelemetry_go_otel//attribute",
     ],
