+++ pkg/kv/bulk/BUILD.bazel
@@ -13,9 +13,9 @@ go_library(
     importpath = "github.com/cockroachdb/cockroach/pkg/kv/bulk",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/ccl/backupccl/backuppb",
         "//pkg/keys",
         "//pkg/kv",
+        "//pkg/kv/bulk/bulkpb",
         "//pkg/kv/kvclient/rangecache",
         "//pkg/kv/kvserver/kvserverbase",
         "//pkg/roachpb",
