+++ pkg/util/bulk/BUILD.bazel
@@ -19,6 +19,7 @@ go_test(
     deps = [
         ":bulk",
         "//pkg/ccl/backupccl/backuppb",
+        "//pkg/kv/bulk/bulkpb",
         "//pkg/roachpb",
         "//pkg/util/tracing",
         "//pkg/util/tracing/tracingpb",
